+++ database/migrations/2022_01_03_031352_create_type_prestations_table.php
@@ -16,7 +16,7 @@ class CreateTypePrestationsTable extends Migration
         Schema::create('type_prestations', function (Blueprint $table) {
             $table->id();
             $table->string('libelle');
-            $table->double('montant')->default(0);
+            $table->decimal('montant', 18, 2)->default(0);
             $table->boolean('delete_ayant_droit')->default(0); //0 pour non et 1 pour oui
             $table->timestamps();
         });
